requirements.txt
openai
spotipy
streamlit
python-dotenv
pydub